backend/requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
anthropic==0.69.0
python-decouple==3.8
dj-database-url==3.0.1
